Conclusão dos DAL, BLL, Controller e revisão de código.

+++ EmergencyManagementSystem.Common.Entities/Entities/Occupation.cs
@@ -1,15 +1,10 @@
 ﻿using EmergencyManagementSystem.Common.Entities.Interfaces;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace EmergencyManagementSystem.Common.Entities.Entities
 {
-    public class Occupation : IEntity<long>
+    public class Occupation : IEntity<short>
     {
-        public long Id { get; set; }
+        public short Id { get; set; }
         public string Profession { get; set; }
     }
 }
